Added `130` - `user-interrupt` to `exitString` and `_code`

+++ bin/build/identical/__returnClean.sh
@@ -10,10 +10,14 @@
 # Delete files or directories and return the same exit code passed in.
 # Argument: exitCode - Required. Integer. Exit code to return.
 # Argument: item - Optional. One or more files or folders to delete, failures are logged to stderr.
-# Requires: isUnsignedInteger _argument __environment
+# Requires: isUnsignedInteger _argument __environment usageDocument
 returnClean() {
   local exitCode="${1-}" && shift
-  isUnsignedInteger "$exitCode" || _argument "${FUNCNAME[0]} $exitCode (not an integer) $*" || return $?
-  __environment rm -rf "$@" || return "$exitCode"
-  return "$exitCode"
+  if ! isUnsignedInteger "$exitCode"; then
+    local this="${FUNCNAME[0]}"
+    [ "$exitCode" = "--help" ] && usageDocument "$this" "${BASH_SOURCE[0]}" 0 || _argument "$this $exitCode (not an integer) $*" || return $?
+  else
+    __environment rm -rf "$@" || return "$exitCode"
+    return "$exitCode"
+  fi
 }
